pyproject.toml
[project]
name = "h"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "google-genai>=1.68.0",
    "google-generativeai>=0.8.6",
    "openai>=2.30.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
